some edits
