+++ tests/unit_tests/test_mesh_validator.py
@@ -10,8 +10,8 @@ class TestMeshValidator(unittest.TestCase):
    
 
    def setUp(self):
-     self.mesh_validator = MeshValidator("../regression_tests/example_meshes/abstract_env_meshes/hgrad.json")
-    #  self.mesh_validator = MeshValidator("./resources/amsr_3day_single_SIC_split_config.json") 
+    #  self.mesh_validator = MeshValidator("../regression_tests/example_meshes/abstract_env_meshes/hgrad.json")
+     self.mesh_validator = MeshValidator("./resources/amsr_3day_single_SIC_split_config.json") 
 
 
    def test_sampler(self):
@@ -29,6 +29,7 @@ class TestMeshValidator(unittest.TestCase):
 
    def test_validate_mesh(self):
       distance = self.mesh_validator.validate_mesh()
+      print (distance)
       self.assertLess (distance, 0.1)
